refactor: Move out TargetActionMenu
